Move FilterHopperBlockEntity file location
